# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "back-end part for mobile application.",
        "title": "Learn-Share API",
        "contact": {
            "name": "Ruslan's Support",
            "url": "https://t.me/Ruslan20007",
            "email": "ruslanrbb8@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:81",
    "basePath": "/api",
    "paths": {
...
        "/lesson": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Check is all data confirmed and if so create lesson with status \"verification\" (Unconfirmed)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lessons"
                ],
                "summary": "Add Unconfirmed lesson (lesson request)",
                "parameters": [
                    {
                        "description": "LessonData",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/book_lesson.request"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    }
                }
            }
        },
        "/lessons/{id}": {
            "get": {
                "description": "Return lesson data by lesson's id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lessons"
                ],
                "summary": "Get lesson data by lesson's id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "LessonID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/get_lesson.response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    }
                }
            }
        },
        "/lessons/{id}/approve": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Set lesson status \"waiting\" if this user is a teacher to lesson and lesson hasn't been cancelled (was verification)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lessons"
                ],
                "summary": "Approve lesson",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "LessonID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    }
                }
            }
        },
...
        "/lessons/{id}/finish": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Set lesson status \"finished\" if this user is a teacher to lesson and lesson's stats has been \"ongoing\"",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lessons"
                ],
                "summary": "Finished lesson",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "LessonID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    }
                }
            }
        },
...
        "/student/lessons": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return all lessons which have student",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "students"
                ],
                "summary": "Get lessons for students",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/get_student_lessons.response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    }
                }
            }
        },
...
        "/teacher/lessons": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return all lessons which have teacher",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teachers"
                ],
                "summary": "Get lessons for teachers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/get_teacher_lessons.response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "book_lesson.request": {
            "description": "book lesson body request",
            "type": "object",
            "required": [
                "category_id",
                "schedule_time_id",
                "teacher_id"
            ],
            "properties": {
                "category_id": {
                    "type": "integer",
                    "example": 1
                },
                "schedule_time_id": {
                    "type": "integer",
                    "example": 1
                },
                "teacher_id": {
                    "description": "@Description exactly teacherID, not his userID",
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "get_lesson.response": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer",
                    "example": 1
                },
                "category_name": {
                    "type": "string",
                    "example": "Programming"
                },
                "datetime": {
                    "type": "string",
                    "example": "2025-02-01T09:00:00Z"
                },
                "lesson_id": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "string",
                    "example": "verification"
                },
                "student_avatar": {
                    "type": "string",
                    "example": "uuid.png"
                },
                "student_email": {
                    "type": "string",
                    "example": "test@test.com"
                },
                "student_id": {
                    "type": "integer",
                    "example": 1
                },
                "student_name": {
                    "type": "string",
                    "example": "John"
                },
                "student_surname": {
                    "type": "string",
                    "example": "Smith"
                },
                "teacher_avatar": {
                    "type": "string",
                    "example": "uuid.png"
                },
                "teacher_email": {
                    "type": "string",
                    "example": "test@test.com"
                },
                "teacher_id": {
                    "type": "integer",
                    "example": 1
                },
                "teacher_name": {
                    "type": "string",
                    "example": "John"
                },
                "teacher_surname": {
                    "type": "string",
                    "example": "Smith"
                },
                "teacher_user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "get_student_lessons.respLessons": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer",
                    "example": 1
                },
                "category_name": {
                    "type": "string",
                    "example": "Programming"
                },
                "datetime": {
                    "type": "string",
                    "example": "2025-02-01T09:00:00Z"
                },
                "lesson_id": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "string",
                    "example": "verification"
                },
                "teacher_avatar": {
                    "type": "string",
                    "example": "uuid.png"
                },
                "teacher_email": {
                    "type": "string",
                    "example": "test@test.com"
                },
                "teacher_id": {
                    "type": "integer",
                    "example": 1
                },
                "teacher_name": {
                    "type": "string",
                    "example": "John"
                },
                "teacher_surname": {
                    "type": "string",
                    "example": "Smith"
                },
                "teacher_user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "get_student_lessons.response": {
            "type": "object",
            "properties": {
                "lessons": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/get_student_lessons.respLessons"
                    }
                }
            }
        },
...
        "get_teacher_lessons.respLessons": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer",
                    "example": 1
                },
                "category_name": {
                    "type": "string",
                    "example": "Programming"
                },
                "datetime": {
                    "type": "string",
                    "example": "2025-02-01T09:00:00Z"
                },
                "lesson_id": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "string",
                    "example": "verification"
                },
                "student_avatar": {
                    "type": "string",
                    "example": "uuid.png"
                },
                "student_email": {
                    "type": "string",
                    "example": "test@test.com"
                },
                "student_id": {
                    "type": "integer",
                    "example": 1
                },
                "student_name": {
                    "type": "string",
                    "example": "John"
                },
                "student_surname": {
                    "type": "string",
                    "example": "Smith"
                }
            }
        },
        "get_teacher_lessons.response": {
            "type": "object",
            "properties": {
                "lessons": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/get_teacher_lessons.respLessons"
                    }
                }
            }
        },
...
        "httputils.ErrorStruct": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}